setup angular plus corrections repository

+++ src/main/java/com/narbaniki/timescrimble/PartieRepository.java
@@ -2,6 +2,6 @@ package com.narbaniki.timescrimble;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PartieRepository extends JpaRepository<Partie, Integer>{
+public interface PartieRepository extends JpaRepository<Partie, Long>{
     
 }
